Use variables and functions to make things clearer
I find it easier to tell when colors of different styles are the same
when variables are used. Additionally, there are some instances where
hand-creating a color (such as the current line highlight) gets in the
way of future enhancements. This change introduces variables for all
the appropriate color settings and uses functions such as lighten()
and contrast() to make keeping the stylesheet up-to-date simpler.